implement multi traveler GUI animation and process synchronization

+++ main_m4.c
@@ -4,12 +4,14 @@
 #include <signal.h>
 #include <sys/wait.h>
 #include <sys/types.h>
+#include "raylib.h"
 #include "Graph.h"
 #include "InputHandler.h"
 #include "Dijkstra.h"
+#include "Visualizer.h"
+#include "Animation.h"
 
 int main() {
-
     Graph *graph = NULL;
     Traveler *travelers = NULL;
     int travelerCount = 0;
@@ -18,71 +20,95 @@ int main() {
         return 1;
     }
 
-    printf("=== Milestone 4: Multi Process System ===\n");
-    printf("Loaded %d travelers successfully.\n\n", travelerCount);
-
     for (int i = 0; i < travelerCount; i++) {
-
-        printf("[Traveler ID: %d] Calculating shortest path from %d to %d:\n",
-               travelers[i].id,
-               travelers[i].src,
-               travelers[i].dst);
-
         dijkstra(graph,
                  travelers[i].src,
                  travelers[i].dst,
                  travelers[i].path,
                  &travelers[i].pathLength,
                  &travelers[i].totalDistance);
-
-        printf("Total Distance: %d\n", travelers[i].totalDistance);
-        printf("-----------------------------------------------\n\n");
     }
 
-
     for (int i = 0; i < travelerCount; i++) {
-
         pid_t pid = fork();
 
         if (pid < 0) {
-
             perror("fork failed");
             return 1;
         }
 
         if (pid == 0) {
-
-            printf("[%d] started\n", getpid());
-
+            printf("[Traveler %d | PID %d] started\n", travelers[i].id, getpid());
             pause();
-
             exit(0);
-
         } else {
-
             travelers[i].pid = pid;
         }
     }
 
-    printf("Parent created all child processes.\n");
+    const int screenWidth = 650;
+    const int screenHeight = 550;
+
+    InitWindow(screenWidth, screenHeight, "Milestone 4 - Multi Traveler Simulation");
+    InitGraphLayout(screenWidth, screenHeight);
+    SetTargetFPS(60);
+
+    MovingEntity entities[100];
+
+    Color travelerColors[6] = {
+        RED, BLUE, GREEN, ORANGE, PURPLE, MAROON
+    };
 
     for (int i = 0; i < travelerCount; i++) {
+        InitEntity(&entities[i], GetNodePosition(travelers[i].path[0]));
+        entities[i].status = ENTITY_MOVING;
+    }
 
-        sleep(2);
+    while (!WindowShouldClose()) {
+        float deltaTime = GetFrameTime();
 
-        printf("Parent terminating traveler PID %d\n",
-               travelers[i].pid);
+        BeginDrawing();
 
-        kill(travelers[i].pid, SIGTERM);
+        DrawRectangleGradientV(0, 0, screenWidth, screenHeight,
+                               (Color){18, 24, 38, 255},
+                               (Color){45, 62, 90, 255});
+
+        drawEdges(graph);
+        DrawStaticGraph();
+
+        for (int i = 0; i < travelerCount; i++) {
+            if (!entities[i].isFinished) {
+                UpdateEntity(&entities[i],
+                             travelers[i].path,
+                             travelers[i].pathLength,
+                             deltaTime,
+                             graph);
+            }
+
+            DrawMovingEntity(entities[i], travelerColors[i % 6]);
+
+            if (entities[i].isFinished && travelers[i].pid > 0) {
+                printf("Traveler %d arrived. Terminating PID %d\n",
+                       travelers[i].id,
+                       travelers[i].pid);
+
+                kill(travelers[i].pid, SIGTERM);
+                waitpid(travelers[i].pid, NULL, 0);
+                travelers[i].pid = -1;
+            }
+        }
+
+        EndDrawing();
     }
 
     for (int i = 0; i < travelerCount; i++) {
-
-        waitpid(travelers[i].pid, NULL, 0);
+        if (travelers[i].pid > 0) {
+            kill(travelers[i].pid, SIGTERM);
+            waitpid(travelers[i].pid, NULL, 0);
+        }
     }
 
-    printf("All child processes finished.\n");
-
+    CloseWindow();
 
     free(travelers);
     freeGraph(graph);
